move creator from problem file to physicsparticles.hpp
