+++ src/particles/PhysicsParticles.hpp
@@ -16,6 +16,7 @@
 #include "AMReX_REAL.H"
 #include "AMReX_SPACE.H"
 #include "AMReX_Vector.H"
+#include "hydro/hydro_system.hpp"
 #include "physics_info.hpp"
 
 // Macro to create bit flags: BITFLAG(position) = 2^(position - 1)
@@ -270,15 +271,13 @@ class PhysicsParticleDescriptorBase
 // Traits class for specializing particle creation behavior
 template <ParticleType particleType> struct ParticleCreationTraits {
 	// Default nested ParticleChecker - determines if a particle should be created at a location
-	template <typename problem_t>
-	struct ParticleChecker {
+	template <typename problem_t> struct ParticleChecker {
 		double param1;
 		double param2;
 		AMREX_GPU_HOST_DEVICE ParticleChecker(double t1, double t2) : param1(t1), param2(t2) {}
 
 		AMREX_GPU_DEVICE auto operator()(amrex::Array4<const amrex::Real> const &state_arr, int i, int j, int k,
-						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dx, 
-						 amrex::Real current_time, amrex::Real dt) const -> bool
+						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dx, amrex::Real current_time, amrex::Real dt) const -> bool
 		{
 			// Default implementation creates no particles
 			amrex::ignore_unused(state_arr, i, j, k, dx, current_time, dt, param1, param2);
@@ -287,8 +286,7 @@ template <ParticleType particleType> struct ParticleCreationTraits {
 	};
 
 	// Default nested ParticleCreator - initializes a particle's properties
-	template <typename problem_t>
-	struct ParticleCreator {
+	template <typename problem_t> struct ParticleCreator {
 		int mass_idx;
 		int cpu_id;
 		amrex::Long pid_start;
@@ -296,18 +294,15 @@ template <ParticleType particleType> struct ParticleCreationTraits {
 		amrex::Real param2;
 
 		AMREX_GPU_HOST_DEVICE
-		ParticleCreator(int mass_index, int processor_id, amrex::Long particle_id_start, 
-				amrex::Real param1, amrex::Real param2)
-			: mass_idx(mass_index), cpu_id(processor_id), pid_start(particle_id_start), 
-			  param1(param1), param2(param2)
+		ParticleCreator(int mass_index, int processor_id, amrex::Long particle_id_start, amrex::Real param1, amrex::Real param2)
+		    : mass_idx(mass_index), cpu_id(processor_id), pid_start(particle_id_start), param1(param1), param2(param2)
 		{
 		}
 
 		template <typename ParticleType, typename StateArray>
 		AMREX_GPU_DEVICE void operator()(ParticleType &p, StateArray const &state_arr, int i, int j, int k,
-						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dx, 
-						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &plo,
-						 amrex::Long particle_offset) const
+						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dx,
+						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &plo, amrex::Long particle_offset) const
 		{
 			// Default implementation does nothing
 			amrex::ignore_unused(p, state_arr, i, j, k, dx, plo, particle_offset);
@@ -316,9 +311,8 @@ template <ParticleType particleType> struct ParticleCreationTraits {
 
 	// Main method to create particles
 	template <typename problem_t, typename ContainerType>
-	static void createParticles(ContainerType *container, int mass_idx, amrex::MultiFab &state, 
-				int lev, amrex::Real current_time, amrex::Real dt,
-				amrex::Real param1, amrex::Real param2)
+	static void createParticles(ContainerType *container, int mass_idx, amrex::MultiFab &state, int lev, amrex::Real current_time, amrex::Real dt,
+				    amrex::Real param1, amrex::Real param2)
 	{
 		// Default implementation does nothing
 	}
@@ -327,41 +321,27 @@ template <ParticleType particleType> struct ParticleCreationTraits {
 // Specialization for CIC particles
 template <> struct ParticleCreationTraits<ParticleType::CIC> {
 	// Specialized nested ParticleChecker for CIC particles
-	template <typename problem_t>
-	struct ParticleChecker {
+	template <typename problem_t> struct ParticleChecker {
 		double param1;
 		double param2;
 		AMREX_GPU_HOST_DEVICE ParticleChecker(double t1, double t2) : param1(t1), param2(t2) {}
 
 		AMREX_GPU_DEVICE auto operator()(amrex::Array4<const amrex::Real> const &state_arr, int i, int j, int k,
-						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dx, 
-						 amrex::Real current_time, amrex::Real dt) const -> bool
+						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dx, amrex::Real current_time, amrex::Real dt) const -> bool
 		{
-			// Default implementation for CIC particles
+			// A simple demonstration of particle creation
 			// Could check density threshold or other state-based conditions
-			amrex::ignore_unused(state_arr);
-			amrex::ignore_unused(i);
-			amrex::ignore_unused(j);
-			amrex::ignore_unused(k);
-			amrex::ignore_unused(dx);
-			amrex::ignore_unused(current_time);
-			amrex::ignore_unused(dt);
-			return false;
-
-			// An example implementation is given below.
-			// const int spacing = 16;
-			// const bool is_create_particle_1 = current_time <= param1 && current_time + dt > param1;
-			// const bool is_create_particle_2 = current_time <= param2 && current_time + dt > param2;
-			// return (is_create_particle_1 || is_create_particle_2) && 
-			//        (i != 0 && i % spacing == 0) && 
-			//        (j != 0 && j % spacing == 0) &&
-			//        (k != 0 && k % spacing == 0);
+			amrex::ignore_unused(state_arr, dx);
+			const int spacing = 16;
+			const bool is_create_particle_1 = current_time <= param1 && current_time + dt > param1;
+			const bool is_create_particle_2 = current_time <= param2 && current_time + dt > param2;
+			return (is_create_particle_1 || is_create_particle_2) && (i != 0 && i % spacing == 0) && (j != 0 && j % spacing == 0) &&
+			       (k != 0 && k % spacing == 0);
 		}
 	};
 
 	// Specialized nested ParticleCreator for CIC particles
-	template <typename problem_t>
-	struct ParticleCreator {
+	template <typename problem_t> struct ParticleCreator {
 		int mass_idx;
 		int cpu_id;
 		amrex::Long pid_start;
@@ -369,60 +349,45 @@ template <> struct ParticleCreationTraits<ParticleType::CIC> {
 		amrex::Real param2;
 
 		AMREX_GPU_HOST_DEVICE
-		ParticleCreator(int mass_index, int processor_id, amrex::Long particle_id_start, 
-				amrex::Real param1, amrex::Real param2)
-			: mass_idx(mass_index), cpu_id(processor_id), pid_start(particle_id_start), 
-			  param1(param1), param2(param2)
+		ParticleCreator(int mass_index, int processor_id, amrex::Long particle_id_start, amrex::Real param1, amrex::Real param2)
+		    : mass_idx(mass_index), cpu_id(processor_id), pid_start(particle_id_start), param1(param1), param2(param2)
 		{
 		}
 
 		template <typename ParticleType, typename StateArray>
 		AMREX_GPU_DEVICE void operator()(ParticleType &p, StateArray const &state_arr, int i, int j, int k,
-						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dx, 
-						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &plo,
-						 amrex::Long particle_offset) const
+						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dx,
+						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &plo, amrex::Long particle_offset) const
 		{
-			// Default implementation for CIC particles
-			amrex::ignore_unused(p);
-			amrex::ignore_unused(state_arr);
-			amrex::ignore_unused(i);
-			amrex::ignore_unused(j);
-			amrex::ignore_unused(k);
-			amrex::ignore_unused(dx);
-			amrex::ignore_unused(plo);
-			amrex::ignore_unused(particle_offset);
+			// A simple demonstration of particle creation
+			p.pos(0) = plo[0] + (i + 0.5) * dx[0];
+			p.pos(1) = plo[1] + (j + 0.5) * dx[1];
+			p.pos(2) = plo[2] + (k + 0.5) * dx[2];
 
-			// An example implementation is given below.
-			// // Set particle position at cell center
-			// p.pos(0) = plo[0] + (i + 0.5) * dx[0];
-			// p.pos(1) = plo[1] + (j + 0.5) * dx[1];
-			// p.pos(2) = plo[2] + (k + 0.5) * dx[2];
+			// Set particle ID and CPU
+			p.id() = pid_start + particle_offset;
+			p.cpu() = cpu_id;
 
-			// // Set particle ID and CPU
-			// p.id() = pid_start + particle_offset;
-			// p.cpu() = cpu_id;
+			// Set particle mass and velocities
+			const amrex::Real cell_volume = AMREX_D_TERM(dx[0], *dx[1], *dx[2]);
+			const amrex::Real cell_density = state_arr(i, j, k, HydroSystem<problem_t>::density_index);
+			const amrex::Real cell_mass = cell_density * cell_volume;
 
-			// // Set particle mass and velocities
-			// const amrex::Real cell_volume = AMREX_D_TERM(dx[0], *dx[1], *dx[2]);
-			// const amrex::Real cell_density = state_arr(i, j, k, HydroSystem<problem_t>::density_index);
-			// const amrex::Real cell_mass = cell_density * cell_volume;
+			// Initialize particle properties
+			p.rdata(mass_idx) = 0.5 * cell_mass;
+			p.rdata(mass_idx + 1) = state_arr(i, j, k, HydroSystem<problem_t>::x1Momentum_index) / cell_density;
+			p.rdata(mass_idx + 2) = state_arr(i, j, k, HydroSystem<problem_t>::x2Momentum_index) / cell_density;
+			p.rdata(mass_idx + 3) = state_arr(i, j, k, HydroSystem<problem_t>::x3Momentum_index) / cell_density;
 
-			// // Initialize particle properties
-			// p.rdata(mass_idx) = 0.5 * cell_mass;
-			// p.rdata(mass_idx + 1) = state_arr(i, j, k, HydroSystem<problem_t>::x1Momentum_index) / cell_density;
-			// p.rdata(mass_idx + 2) = state_arr(i, j, k, HydroSystem<problem_t>::x2Momentum_index) / cell_density;
-			// p.rdata(mass_idx + 3) = state_arr(i, j, k, HydroSystem<problem_t>::x3Momentum_index) / cell_density;
-
-			// // Update cell density (remove mass that was given to particle)
-			// state_arr(i, j, k, HydroSystem<problem_t>::density_index) = 0.5 * cell_density;
+			// Update cell density (remove mass that was given to particle)
+			state_arr(i, j, k, HydroSystem<problem_t>::density_index) = 0.5 * cell_density;
 		}
 	};
 
 	// Main method to create particles
 	template <typename problem_t, typename ContainerType>
-	static void createParticles(ContainerType *container, int mass_idx, amrex::MultiFab &state, 
-				int lev, amrex::Real current_time, amrex::Real dt,
-				amrex::Real param1, amrex::Real param2)
+	static void createParticles(ContainerType *container, int mass_idx, amrex::MultiFab &state, int lev, amrex::Real current_time, amrex::Real dt,
+				    amrex::Real param1, amrex::Real param2)
 	{
 		if (container != nullptr) {
 			if (mass_idx >= 0 && mass_idx + 3 < ContainerType::ParticleType::NReal) {
@@ -476,8 +441,8 @@ template <> struct ParticleCreationTraits<ParticleType::CIC> {
 						const amrex::IntVect iv(AMREX_D_DECL(i, j, k));
 						const auto index = box.index(iv);
 
-						if (pcounts[index] > 0) {                                // NOLINT
-							auto &p = pdata[poffset[index]];                     // NOLINT
+						if (pcounts[index] > 0) {						  // NOLINT
+							auto &p = pdata[poffset[index]];				  // NOLINT
 							particle_creator(p, state_arr, i, j, k, dx, plo, poffset[index]); // NOLINT
 						}
 					});
