+++ src/problems/Gravity3D/gravity_3d.cpp
@@ -63,73 +63,6 @@ template <> struct Physics_Traits<BinaryOrbit> {
 	static constexpr double radiation_constant = 1.0;
 };
 
-namespace quokka
-{
-// Specialization of ParticleCreationTraits for BinaryOrbit
-template <> struct ParticleCreationTraits<ParticleType::CIC>::ParticleChecker<BinaryOrbit> {
-	double param1;
-	double param2;
-	AMREX_GPU_HOST_DEVICE ParticleChecker(double t1, double t2) : param1(t1), param2(t2) {}
-
-	AMREX_GPU_DEVICE auto operator()(array_t const & /*state_arr*/, int i, int j, int k, 
-                                     amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const & /*dx*/,
-					                 amrex::Real current_time, amrex::Real dt) const -> bool
-	{
-		const int spacing = 16;
-		const bool is_create_particle_1 = current_time <= param1 && current_time + dt > param1;
-		const bool is_create_particle_2 = current_time <= param2 && current_time + dt > param2;
-		return (is_create_particle_1 || is_create_particle_2) && 
-               (i != 0 && i % spacing == 0) && 
-               (j != 0 && j % spacing == 0) &&
-		       (k != 0 && k % spacing == 0);
-	}
-};
-
-// Specialization of ParticleCreator for BinaryOrbit
-template <> struct ParticleCreationTraits<ParticleType::CIC>::ParticleCreator<BinaryOrbit> {
-	int mass_idx;
-	int cpu_id;
-	amrex::Long pid_start;
-	amrex::Real param1;
-	amrex::Real param2;
-
-	AMREX_GPU_HOST_DEVICE
-	ParticleCreator(int mass_index, int processor_id, amrex::Long particle_id_start, amrex::Real param1, amrex::Real param2)
-	    : mass_idx(mass_index), cpu_id(processor_id), pid_start(particle_id_start), param1(param1), param2(param2)
-	{
-	}
-
-	template <typename ParticleType, typename StateArray>
-	AMREX_GPU_DEVICE void operator()(ParticleType &p, StateArray const &state_arr, int i, int j, int k,
-					 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dx, amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &plo,
-					 amrex::Long particle_offset) const
-	{
-		// Set particle position at cell center
-		p.pos(0) = plo[0] + (i + 0.5) * dx[0];
-		p.pos(1) = plo[1] + (j + 0.5) * dx[1];
-		p.pos(2) = plo[2] + (k + 0.5) * dx[2];
-
-		// Set particle ID and CPU
-		p.id() = pid_start + particle_offset;
-		p.cpu() = cpu_id;
-
-		// Set particle mass and velocities
-		const amrex::Real cell_volume = AMREX_D_TERM(dx[0], *dx[1], *dx[2]);
-		const amrex::Real cell_density = state_arr(i, j, k, HydroSystem<BinaryOrbit>::density_index);
-		const amrex::Real cell_mass = cell_density * cell_volume;
-
-		// Initialize particle properties
-		p.rdata(mass_idx) = 0.5 * cell_mass;
-		p.rdata(mass_idx + 1) = state_arr(i, j, k, HydroSystem<BinaryOrbit>::x1Momentum_index) / cell_density;
-		p.rdata(mass_idx + 2) = state_arr(i, j, k, HydroSystem<BinaryOrbit>::x2Momentum_index) / cell_density;
-		p.rdata(mass_idx + 3) = state_arr(i, j, k, HydroSystem<BinaryOrbit>::x3Momentum_index) / cell_density;
-
-		// Update cell density (remove mass that was given to particle)
-		state_arr(i, j, k, HydroSystem<BinaryOrbit>::density_index) = 0.5 * cell_density;
-	}
-};
-} // namespace quokka
-
 template <> void QuokkaSimulation<BinaryOrbit>::setInitialConditionsOnGrid(quokka::grid const &grid_elem)
 {
 	const amrex::Box &indexRange = grid_elem.indexRange_;
